LPS-58771 Move getTemplateBundle to tracker
